[#20] ♻️ refactor : Auth User 합치기
