+++ src/main/java/com/sparta/hotitemcollector/domain/user/UserService.java
@@ -1,15 +1,13 @@
-package com.sparta.hotitemcollector.domain.user.service;
+package com.sparta.hotitemcollector.domain.user;
 
+import com.sparta.hotitemcollector.domain.security.UserDetailsImpl;
 import com.sparta.hotitemcollector.domain.token.Token;
 import com.sparta.hotitemcollector.domain.token.TokenService;
-import com.sparta.hotitemcollector.domain.user.User;
-import com.sparta.hotitemcollector.domain.user.UserRepository;
-import com.sparta.hotitemcollector.domain.user.UserRole;
-import com.sparta.hotitemcollector.domain.user.UserStatus;
 import com.sparta.hotitemcollector.domain.user.dto.auth.LoginReqeustDto;
 import com.sparta.hotitemcollector.domain.user.dto.auth.LoginResponseDto;
 import com.sparta.hotitemcollector.domain.user.dto.auth.RefreshRequestDto;
 import com.sparta.hotitemcollector.domain.user.dto.auth.SignupRequestDto;
+import com.sparta.hotitemcollector.domain.user.dto.user.*;
 import com.sparta.hotitemcollector.global.exception.CustomException;
 import com.sparta.hotitemcollector.global.exception.ErrorCode;
 import com.sparta.hotitemcollector.global.jwt.JwtUtil;
@@ -19,13 +17,13 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
 import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
 @Slf4j(topic = "UserService 로그")
-public class AuthService {
-
+public class UserService {
     private final JwtUtil jwtUtil;
 
     private final PasswordEncoder passwordEncoder;
@@ -170,6 +168,69 @@ public class AuthService {
     }
 
 
+
+
+    public ProfileResponseDto updateProfile(ProfileRequestDto requestDto, User user) {
+        User findUser = userRepository.findById(user.getId()).orElseThrow(()->new CustomException(ErrorCode.NOT_FOUND_USER));
+
+        // 유저 정보 업데이트 (요청된 필드만 업데이트)
+        if (requestDto.getNickname() != null) {
+            findUser.setNickname(requestDto.getNickname());
+        }
+        if (requestDto.getPhoneNumber() != null) {
+            findUser.setPhoneNumber(requestDto.getPhoneNumber());
+        }
+        if (requestDto.getAddress() != null) {
+            findUser.setAddress(requestDto.getAddress());
+        }
+        if (requestDto.getInfo() != null) {
+            findUser.setInfo(requestDto.getInfo());
+        }
+        if(requestDto.getProfileImage() != null){
+            findUser.setProfileImage(requestDto.getProfileImage());
+        }
+        // 유저 저장
+        User saveUser = userRepository.save(findUser);
+        return new ProfileResponseDto(saveUser);
+    }
+
+    public void updatePassword(updatePasswordRequestDto requestDto, User user) {
+        // 유저를 찾고, 없으면 예외를 발생시킵니다.
+        User findUser = userRepository.findById(user.getId())
+                .orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND_USER));
+
+        // 기존 비밀번호가 맞는지 확인합니다.
+        if (!passwordEncoder.matches(requestDto.getOldPassword(), findUser.getPassword())) {
+            throw new CustomException(ErrorCode.INCORRECT_PASSWORD);
+        }
+
+        // 새로운 비밀번호 설정
+        findUser.setPassword(passwordEncoder.encode(requestDto.getNewPassword()));
+
+        // 변경된 비밀번호를 저장합니다.
+        userRepository.save(findUser);
+    }
+
+    public GetUserProfileDto getUserProfile(Long userId, Optional<UserDetailsImpl> currentUser) {
+        User findUser = userRepository.findById(userId)
+                .orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND_USER));
+        if(!currentUser.isEmpty()){
+            boolean isOwnProfile = findUser.getId().equals(currentUser.get().getUser().getId());
+
+            return new GetUserProfileDto(findUser, isOwnProfile);
+        }else{
+            return new GetUserProfileDto(findUser, false);
+        }
+
+    }
+
+
+    public UserAddressDto getUserAddress(User user) {
+        User findUser = userRepository.findById(user.getId())
+                .orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND_USER));
+        return new UserAddressDto(findUser);
+    }
+
     /**
      * username 을 통해 유저 정보 가져오기
      *
@@ -191,6 +252,17 @@ public class AuthService {
         if (!user.isExist()) {
             throw new CustomException(ErrorCode.WITHDRAW_USER);
         }
+    }
+
+
+
+    /**
+     * 닉네임으로 유저 리스트 검색
+     * @param nickname
+     * @return
+     */
+    public List<User> findByNicknameContainingIgnoreCase (String nickname) {
+        return userRepository.findByNicknameContainingIgnoreCase(nickname);
     }
